rasters of global heatwave duration per decade

a raster with layers by decade

Usage

metrics_by_decade_raster(
  mhwdb_conn,
  mhw_table,
  mhw_metric = "int_mean",
  sql_function = "avg",
  decades = c("2040", "2050", "2060"),
  ensemble_list_string = NA,
  crs = "EPSG:4087"
)

Arguments

mhwdb_conn

database connection

mhw_table

character string name of the table to query, required (e.g. arise10_metrics)

ensemble_list_string

character string that is a list of ensembles to include in format "006,007,008" ensembles have leaving zeros, are 3 digits and this must be a separated list with leading zeros

crs

character coordinate reference system, EPSG:4087, alternatives are EPSG:4326.

Value

terra raster stack of layers for each decade