find value for percentile cut-off

Usage

percentile_cutoff_value(x, cut_percent = 1)

Arguments

x

values to find the cut-off in, numeric vector

cut_percent

numeric top percent to cut e.g. 1 to cut top 1

numeric value in x that is the max befoe the percentile indicated by cut_percent given a range of values, find the max value just before a percentile, say 1 used to divide a range into two bins or set an upper limit. If a whole number is sent, the percentiles are sequenced by 1 like 2.25