create raster of global marine heatwave summary metric per location

using SQL, for example generated by mhw_metric_summary_sql, creates a single raster version with lat, long and metric The raster will have attributes lat, lon and the name of the metric sent

Usage

summary_metrics_raster(
  mhwdb_conn,
  mhw_sql,
  mhw_value_name,
  rotate_globe = TRUE,
  crs = "EPSG:4087"
)

Arguments

mhwdb_conn

database connection

mhw_sql

character string of SQL that calculates summary. SQL must output table with 3 colums (lat, lon, some_column)

rotate_globe

= TRUE boolean True mean rotate the global output so prime meridian is centered, False, no rotation applied, this works if data is maped from 0 to 360 instead of -180 to 180

crs

= 'EPSG:4087' optional character coordinate reference system

Value

single terra raster, if